requirements.txt
numpy>=1.24
scipy>=1.10
matplotlib>=3.7
PyYAML>=6.0
Shapely>=2.0
plotly>=5.18
